requirements.txt
appdirs==1.4.4
asgiref==3.7.2
beautifulsoup4==4.12.3
black==23.7.0
blinker==1.6.2
blueprint==3.4.2
boto3==1.34.48
botocore==1.34.48
cachelib==0.10.2
certifi==2024.2.2
click==8.1.7
dnspython==2.5.0
email-validator==2.1.0.post1
Flask==2.3.3
Flask-Mail==0.9.1
Flask-Session==0.5.0
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonify==0.5
lxml==5.1.0
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pyee==11.1.0
pyppeteer==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.0
s3transfer==0.10.0
six==1.16.0
soupsieve==2.5
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==1.26.18
websockets==10.4
Werkzeug==2.3.7
zipp==3.17.0
